requirements.txt
playwright==1.51.0
pytest==8.3.5
pytest_bdd==8.1.0
python-dotenv==1.1.0
